updates to iot contract v4.1
// v4.1   KL May 27 2016 Store transaction UUID and timestamp in the
state, aligning world state with the transaction
//
(and by definition the block) that is responsible for this specific
asset state. Note: Initial
//                       commit of this
version was at 4.0 and was in the trade lane folder. Subsequently moved
to the iot
//                       folder and updated to 4.1. Also:
//
- Additional feature in 4.1 -- issue #4 -- rules engine must return
err
//                         - Tweaks to deletePropertiesFromAsset (a
bug found in the alert boilerplate using args instead of ledger)
//
- Added postman tests for delete properties from asset and validation
test rules
//                         - Reordered and normalized the
rules engine call to make original event available and have
consistency
//                           between create, update and
delete properties.