propagate new schema checker / generator
The new schema checker / generator in the aviation sample generates
identical schemas in the iot and trade lane folders and so becomes the
new version for all. Also removed reference to the non-existent
"readContractObjectModel" API in the schemas for iot and trade lane as
the new checker throws an error.