restructure advanced contracts, update README
The advanced contracts iot and aviation will evolve to become an sdk and
so are being shuffled under a new folder of that name. The docs folder
is hoisted from iot to become the only docs folder that evolves going
forward. The README now documents the various compatibility levels of
folders in the project.