reinitialize after clearing world state
When clearing world state, the chaincode must normally be killed,
restarted, and redeployed in order to reinitialize world state. This
commit adds inline reinitialization with the original parameters from
contract state. Note that I read contract state AFTER world state has
been purged. This works because the update is only committed after the
function returns. We'll leave it like this so that we are aprised
immediately of changes in that behavior.