improve query error handling on monitoring ui
The monitoring UI was unable to display errors from some of the
contracts because is uses the snack bar for all errors. The snackbar is
simply not big enough for that and much of the error message is lost. I
moved QUERY errors into a new entry in the middle column so the entire
message can be read. In cases where the query was for an asset that has
not yet been created, the user can leave the entry there and clock
polling. This has the advantage of immediately displaying the new entry
when it is created. Also, this closes an obvious case where the feedback
was just not obvious. An error in your face is the way to go when
debugging chaincode.