handle missing entries much better in aviation contract
There is an issue in hyperledger 0.5 where reading a non-existent key
will succeed and leave you with no errors and no data (byte array is
length zero.) This commit fixes that issue in the common asset layer, so
it works for all of them. There was also some confusion in the test for
error and length in assetCommon that has been fixed.