adapt the assetID mechanism from the aviation contract
Reduce memory consumption and disk usage by removing the large asset
tracking array
from contract state. This required changes to "assetIsActive" and to
delete all and read all
asset code to use an iterator instead of the array. Note that there is
now a dependency on
all assets beginning with a prefix "IOT" ... basic featuresd work
without it, but read all and
delete all rely on it. The next generation contract will formalize this
as a hidden configuration
feature. But for long term tests, we will have to live with this
constraint for a few weeks.