add error and success events for deploy and invoke
For v0.6 fabric, the block formats have been reduced in functionality
significantly. Chaincode errors, which should always be deterministic,
have been removed along with non-deterministic errors such as network
errors. This results in the entire transactions section missing
sometimes, and always guarantees that the block can no longer be used
to diagnose chain code problems. This commit adds both success and
failure events to every invoke and deploy so that the blocks once again
become useful as debugging tools. Also, the audit trail is once again
intact with this change.

This feature should be adjusted if the fabric eventually adds the
errors back, and preferably adds a system defined chain code result
event.