made decode of block txn payload reliable
Function and Args were not easily parsed from a transaction payload in
a block because the separators are random strings of non-printable
characters. The original algorithm had been using split(\n) and that
was not reliable. I tried splitting on strings of non-printable
characters, but the position in the array was random, even for
identical calls. So …
— add func to loop over schema API names and index into payload after
base64 decode, return if found
— for args, look for first index of ‘{‘ and last index of ‘}’ and if
the positions are right, select that string.

Has proven reliable on several contracts in Bluemix and in debug.