update filtered queries for container and minimal sample  POSTMAN tests
These had not been updated when the restriction to filter only asset
state was lifted. Any of the major object attached to the asset state
struct can be searched, so to filter by the asset's state, you have to
prefix it with "assetstate". A classic test for container is
"assetstate.contain.carrier" with the value of the carrier, e.g. "UPS".